move filegenerator into utils and create cli.py
