+++ generation_utils/__init__.py
@@ -5,3 +5,4 @@ from .format_precice_config import PrettyPrinter
 from .other_files_generator import OtherFilesGenerator
 from .config_generator import ConfigGenerator
 from .readme_generator import ReadmeGenerator
+from .file_generator import FileGenerator
